Driver added to permission.

+++ app/Http/Controllers/Dashboard/Customers/Controller.php
@@ -8,6 +8,8 @@ use App\Models\Customer;
 
 use Illuminate\Http\JsonResponse;
 
+use Illuminate\Support\Facades\Gate;
+
 use App\Http\Controllers\Controller as Controllers;
 
 use App\Logic\Dashboard\CRUD\Requests\Customers as CustomersRequest;
@@ -49,6 +51,12 @@ final class Controller extends Controllers
      */
     public function index(CustomersRequest $request): JsonResponse
     {
+        if(! Gate::check('Customers')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'filters' => $this->service->getAllFilters($request),
 
@@ -63,6 +71,12 @@ final class Controller extends Controllers
      */
     public function store(CustomersRequest $request): JsonResponse
     {
+        if(! Gate::check('Customers')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'message' => 'The customer was successfully created.',
 
@@ -77,6 +91,12 @@ final class Controller extends Controllers
      */
     public function show(Customer $customer): JsonResponse
     {
+        if(! Gate::check('Customers')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'customer' => $this->service->showCustomer($customer),
         ]);
@@ -91,6 +111,12 @@ final class Controller extends Controllers
      */
     public function update(CustomersRequest $request, Customer $customer): JsonResponse
     {
+        if(! Gate::check('Customers')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'message' => 'The customer was successfully updated.',
 
@@ -105,6 +131,12 @@ final class Controller extends Controllers
      */
     public function destroy($id): JsonResponse
     {
+        if(! Gate::check('Customers')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         $id = $this->service->deleteCustomer($id);
 
         if(!$id){
